SCRD-605 Added ModelPicker Prototype Page
This checkin includes:
Added a model picker page mimicing the mock with some stylings
Added some buttons with stylings
Added templates and mid-scale-kvm-vsa into api.json

On the model picker page:
Be able get templates and present the subset of the template names.
Be able to select the model name and its responsding details.
Be able to move forward and backward