Add backend service
Add a "shim" backend service which, for the most part, delegates calls
to the appropriate back end (mostly ardana).   Also has APIs for
communicating with SUSE Manager and HPE OpenView.

In its default configuration, all backend calls are routed to the json
server.  Details of the API are in server/README.rst