Use real model in CloudModelSummary (#68)
Modify the CloudModelSummary page to use and save the data from the
input model.  To keep with the paradigm of the other pages, the
Update button was removed, and the information is automatically saved
when the Next button is pressed.  Prevents the user from changing
the number to an empty value.