Persist play id state (#94)
Include the deployment play id in the overall application state,
including persisting to and restoring from the progress.  Modified
top-level state handling to be more generic so that additional overall
state variables can more easily be included.  Manage visibility of
next/back buttons in the deployment progress page based on deployment
state Immediately persist the state when first loading if the state is
being reset or being loaded the first time.