Pass node list to os install playbook
Build and pass a payload the the os install playbook to indicate which
nodes should have the operating system installed.  Add REST function in
the server to return list of local ip addresses, which permits the UI to
exclude the local host (the deployer) from the list passed to the
playbook.