Refactor ServerRoleSummary
This is the first in a series of commits that will create subdirectories
under pages for each top-level page (as needed), which will files for
those 'component' that are really specific to the page.  The CSS (.less)
files are also put into a corresponding pages subdir, using the same
name as the javascript code.