Add new shared Dropdown component (#198)
* Add new shared Dropdown component

The ServerDropdown component was originally designed for the ServerUtils
page but has been used in a number of places throughout the code base.
ServerDropdown has a couple of nice features, such as the ability to add
an extra row as an 'empty' option (when no option has yet been
selected), and it uses css classes that are styled to be consistent with
the rest of the application.  But ServerDropdown suffers from a number
of problems including:
- Non-standard api: the properties that you need to send to
  ServerDropdown are entirely different from the standard ones for a
  <select> component
- It supports both a 'default' and 'empty' option, which both attempt to
   provide the same feature in different ways
- It does not support building the option list and sending it as a child
   property (like <select> does)
- It attempts to manage the state (and props) of the component.  Since
  the parent component still has to track the state of the component,
  this extra level of tracking is unecessary.  Worse yet, this extra
  tracking logic interferes with, and sometimes breaks the overall
  logic.

Since the API of this component is different from ServerDropdown, this
component was added to the code base, rather than removing
ServerDropdown and cause a significant disruptive change.
ServerDropdown can be phased out and removed opportunistically.

* Add missing import

* Add example usage

* Explicitly choose the emptyOption when props.value is undefined

The underlying <SELECT> does not adjust its selection when the value
field (which comes from props.value) is undefined.  This is problematic
when the dropdown needs to be forceable reset to its default state.