Update NIC Mapping tab style and interaction
Use a garbage can icon in the main table.  Permit the NIC mapping name
to be modified.  Improve validation so that the Save and - and + icons
are disabled or removed when they are shouldn't be used, such as when
trying add a new row when the current one is invalid.