SCRD-1895 Fix save function on Add Server Manually modal
The server group and NIC mapping fields in the Add Server Manually modal
did not save the default values correctly. This was because the getServerGroups
and getNicMappings functions used to return a list of name strings but was
refactored to return a list of objects, which caused the setting of initial
values for server group and NIC mapping fields to invalid values.