Reload model after editing model files
When model files are edited directly in the validate config page, the
model should be reloaded from disk so that if a user goes back to a
previous page, those direct edits are reflected in the UI.

To help alleviate the problem of saving an improperly formatted yaml
file and failing to reload it, the textarea uses a yaml validator to
detect invalid yaml and avoid saving.

To improve code re-use, both file editors on this page (one on the Model
tab and the other on the Templates and Services tag) now use the
ServerInput component and use common styling for the editor and page
layout.