SCRD-1710 remove green arrow from Networks tab (#236)
* SCRD-1710 remove green arrow from Networks tab

Most items in this story were already taken care of by
previous changes. This change only removes the green arrow
at Add Network action on Networks tab, to be consistent
with the behavior on other tabs. The green arrow implementation
could be resurrected later on with SCRD-1996.

Also includes the following miscellaneous fixes:
- fix lint complaint in ServerRoleSummary.js
- remove extra ';' on Review Configuration Files page